aider: Ensure `setShowModal(false)` is called before `onClose()` in `handleTakeBreak` to hide `showModal` immediately when the button is clicked.
# Aider chat conversation:

USER: fix the button handleTakeBreak need to click twice, the showModal hide, must click the button handleTakeBreak one time and direct to hide showModal

ASSISTANT: To fix the issue where the `handleTakeBreak` button needs to be clicked twice to hide the `showModal`, we need to ensure that the `setShowModal(false)` is called before `onClose()` in the `handleTakeBreak` function. This will ensure that the modal is hidden immediately when the button is clicked.

Here are the *SEARCH/REPLACE* blocks:

src/app/hooks/useTimer.tsx
```typescript
<<<<<<< SEARCH
    onClose();
    setShowModal(false);
=======
    setShowModal(false);
    onClose();
>>>>>>> REPLACE
```

This change will ensure that the modal is hidden immediately when the `handleTakeBreak` button is clicked.